# Some pre-allocation optimizations for accordion, radio_group and fyne_demo