### Description:
<!-- A summary of the change included and which issue it addresses.
Please include any relevant motivation and background. -->

This applies the per-allocation optimization to a few more places. There are probably a few more that could be added but this is a good step forward in the meantime, fixing the more straight-forward cases. In summary, it boils down to:
- Setting the objects in accordion went from (a total time over three calls) `10.022µs` to `1.567µs` (6.4x faster).
- Adding all objects in CreateRenderer of radio_group went from (a total time over four calls) `2.327µs` to `1.486µs` (1.5x faster).
- Creating the icon strings in fyne_demo went from `11.444µs` to `3.566µs` (3.2x faster). 

### Checklist:
<!-- Please tick these as appropriate using [x] -->
- [ ] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.